move serde to openapi directory
