+++ src/PowerSystemsInvestmentsPortfolios.jl
@@ -176,7 +176,8 @@ include("models/cost_functions/investment_cost.jl")
 include("models/cost_functions/CapitalCost.jl")
 include("models/cost_functions/StorageCapitalCost.jl")
 include("openapi/refs.jl")
-include("openapi/converters.jl")
+include("openapi/cost_conversion.jl")
+include("openapi/export_cost_conversion.jl")
 include("models/generated/includes.jl")
 include("investment_schedule.jl")
 
@@ -185,10 +186,11 @@ include("units/conversions.jl")
 include("units/function_conversions.jl")
 
 include("portfolio.jl")
-include("openapi/document.jl")
+include("openapi/import_document.jl")
 include("validation.jl")
 include("time_mapping.jl")
-include("serialization.jl")
+include("openapi/export_document.jl")
+include("openapi/file_io.jl")
 include("utils/getters.jl")
 include("db_parser.jl")
 include("utils/generate_structs.jl")
